Move the code from Infrastructure to Legacy
In doing so, the namespaces have also been updated. The Resources have
also been moved to the app resources folder.
