+++ app/Resources/schemas/ws-authorization.xsd
