+++ app/Resources/schemas/ws-federation.xsd
