+++ app/Resources/schemas/ws-securitypolicy-1.2.xsd
